+++ game/include/game/core/GameSettingsService.h
@@ -15,9 +15,9 @@ public:                                                       \
     ##type get_##name() const { return m_##name; }
 class GameSettingsService
 {
-    GAME_SETTINGS_FIELD(int, paddle_base_speed)
+    GAME_SETTINGS_FIELD(float, paddle_base_speed)
     GAME_SETTINGS_FIELD(game::Size, paddle_base_size)
-    GAME_SETTINGS_FIELD(int, ball_base_speed)
+    GAME_SETTINGS_FIELD(float, ball_base_speed)
     GAME_SETTINGS_FIELD(game::Size, ball_base_size)
 };
